Frontend/requirements.txt
dash==3.0.1
pandas==2.2.3
python-dotenv==1.0.0